pyproject.toml
[project]
name = "langgraph-demo"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.4.8",
    "langchain>=0.1.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["langgraph_demo"]
